Update servicenow-integration.md
In order to be able to handle newlines and other special characters in the mesage we have to base64 endode the message an decode it on Tray.IO again. Also the data we send to tray has to be json, so we're adding the appropriate headers so we do not have to parse the body on tray anymore.